+++ tests/Console/Service/StatsServiceTest.php
@@ -74,6 +74,7 @@ class StatsServiceTest extends TestCase
         'rusage-utime',
         'rusage-stime',
         'uptime',
+        'draining',
         'id',
         'hostname',
     ];